requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
google-generativeai>=0.3.0 